Can Delete Post
Can now delete posts as well

+++ controllers/commentController.js
@@ -50,8 +50,15 @@ exports.removeComment = async (req, res) => {
 
   const client = await pool.connect();
   try {
-    await client.query(`DELETE FROM Comment WHERE id=${commentId}`);
-    req.flash(`success`, `Success! Your comment was deleted.`);
+    const comment = await client.query(
+      `SELECT deviceId from Comment WHERE id=${commentId}`
+    );
+    if (deviceId == comment.rows[0].deviceid) {
+      await client.query(`DELETE FROM Comment WHERE id=${commentId}`);
+      req.flash(`success`, `Success! Your comment was deleted.`);
+    } else {
+      throw new Error(`Client wasn't properly authenticated`);
+    }
   } catch (err) {
     req.flash(
       `error`,
